♻️ Move options.go to public and add aliases
