+++ public/options_test.go
@@ -1,4 +1,4 @@
-package sqlslog
+package public
 
 import "testing"
